load_unit: Add support for non-idempotent loads
This adds the feature to check loads for being to "non-idempotent"
memory regions. In this case we can't speculatively issue loads as we
would potentially destroy information.

We do so by adding three more states to the load FSM:
- `WAIT_UNTIL_NON_SPECULATIVE`: Which defers the load until the commit
stage signals that the load is the next instruction which would be
committed.
- `REPLAY_LOAD`: Which replays the virtual address part.
- `REPLAY_LOAD_SEND_TAG`: Which replays the tag part.

When replaying we can't get any exception (as the whole point is that we
are doing it because we want to do it non-speculative). An assertion make sure that
assumption holds.